+++ mod/main.lua
@@ -6,9 +6,6 @@
 --[[
 
 TODO:
-- USE QUEUE ITEM FOR CHECKPOINT
-- USE QUEUE ITEM FOR SCHOOLBAG
-- change trinket bans to trinket pool API
 - Implement time offsets, show on the first room of each floor
 - Opponent's shadows