+++ HISTORY.md
@@ -1,5 +1,11 @@
 # Racing+ Version History and News
 
+### *v0.14.20* - December 16th
+
+* You can now perform a Schoolbag switch during an item pickup animation, as long as the item is not an active item.
+* Fixed the bug where you could go the wrong way and still complete a multi-character speedrun custom challenge.
+* Fixed the bug where the automatic update was not working properly. You may need to manually download and install this version.
+
 ### *v0.14.14* - December 14th
 
 * The macOS version should actually work now.